docker/configs/python/requirements.txt
Flask
mysql-connector-python
flask_sqlalchemy